requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
distro==1.9.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-google-genai==2.0.7
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.4
langsmith==0.2.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.21.0
networkx==3.4.2
numpy==2.2.1
openai==1.59.3
orjson==3.10.13
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.0
PyMuPDF==1.25.1
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.15.0
sentence-transformers==3.3.1
setuptools==75.7.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.47.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
watchdog==6.0.0
yarl==1.18.3
